+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to semantic versioning for configuration changes.
 
 ## [Unreleased]
 
+### Auto-updated - 2026-01-05 12:39:03
+- Added: .gitignore
+- Modified: scripts/git-autopush-no-html.sh
+- Modified: scripts/git-autopush.sh
+
+## [Unreleased]
+
 ### Auto-updated - 2026-01-05 12:30:01
 - Modified: scripts/net2_user_activity.py
 - Modified: scripts/net2_user_activity_daemon.py